fix: Move `no-unused-vars` override to index.html
Sometimes libraries forces us to declare function params
that we don't really need. In that scenario, being explicit
and reminding the reader that additional variables are
available, can improve the quality of the code.

The rule was already defined in `tests.js`
but now it is defined in the main configuration file.