fix: Move ESLint plugins to peer dependencies (#39)
* fix: Move ESLint plugins to peer dependencies

This is necessary because package managers do not guarantee that they
will install dependencies of dependencies in `node_modules`. The pnpm
package manager does this by default for all packages; Yarn and npm only
sometimes for version conflicts or optimisation. When this happens,
ESLint cannot load the plugin.

To ensure everything works as expected, plugins (and parsers) must be
explicitly added to your project's dev dependencies.

This is kind of a breaking change -- when you update from 2.0.2 it won't
work anymore without installing the plugins -- but we've only just
released 2.0.0 and haven't adopted it widely yet, so I'm gonna make this
a patch release.

* Add `peerDependenciesMeta` section

All peer deps except for `eslint` are optional.